# docs/api-specification.yaml
openapi: 3.1.0
info:
  title: AI Part Designer API
  description: |
    RESTful API for AI-powered 3D part design and generation platform.
    
    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the 
    `Authorization` header as `Bearer <token>`.
    
    ## Rate Limiting
    - Free tier: 30 requests/minute
    - Pro tier: 100 requests/minute
    - Enterprise tier: 500 requests/minute
    
    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in window
    - `X-RateLimit-Reset`: Unix timestamp when window resets
    
    ## Versioning
    API version is specified in the URL path: `/api/v1/`
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@aipartdesigner.com
  license:
    name: Proprietary
    url: https://aipartdesigner.com/terms
servers:
  - url: https://api.aipartdesigner.com/api/v1
    description: Production server
  - url: https://api.staging.aipartdesigner.com/api/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Local development
...
paths:
...
  /designs/{designId}:
    get:
      tags: [Designs]
      summary: Get design details
      operationId: getDesign
      security:
        - bearerAuth: []
      parameters:
        - name: designId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Design details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Design'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    
    patch:
      tags: [Designs]
      summary: Update design metadata
      operationId: updateDesign
      security:
        - bearerAuth: []
      parameters:
        - name: designId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 255
                description:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                isPublic:
                  type: boolean
      responses:
        '200':
          description: Design updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Design'
    
    delete:
      tags: [Designs]
      summary: Delete a design (soft delete)
      operationId: deleteDesign
      security:
        - bearerAuth: []
      parameters:
        - name: designId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Design deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Design:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        templateId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        description:
          type: string
        sourceType:
          type: string
          enum: [template, ai_generated, imported, modified]
        status:
          type: string
          enum: [draft, processing, ready, failed, archived]
        currentVersion:
          $ref: '#/components/schemas/DesignVersion'
        metadata:
          type: object
          properties:
            parameters:
              type: object
            aiPrompt:
              type: string
            dimensions:
              type: object
              properties:
                x:
                  type: number
                y:
                  type: number
                z:
                  type: number
                unit:
                  type: string
            volume:
              type: number
            surfaceArea:
              type: number
            isPrintable:
              type: boolean
        tags:
          type: array
          items:
            type: string
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DesignVersion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        versionNumber:
          type: integer
        fileUrl:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
        fileFormats:
          type: object
          additionalProperties:
            type: string
            format: uri
        parameters:
          type: object
        geometryInfo:
          type: object
          properties:
            boundingBox:
              type: object
              properties:
                x:
                  type: number
                y:
                  type: number
                z:
                  type: number
            volume:
              type: number
            surfaceArea:
              type: number
            triangleCount:
              type: integer
            isManifold:
              type: boolean
        changeDescription:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            requestId:
              type: string
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: You don't have permission to access this resource
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
...
security:
  - bearerAuth: []